+++ js/hitsuji_ranking.js
@@ -1,26 +1,5 @@
 import BackButton from "./back_button.js";
-
-const API_URL = "http://13.231.182.101";
-
-const getRanks = async () => {
-  const response = await fetch(`${API_URL}/api/time_limits`);
-  const data = await response.json();
-  const ranks = [];
-  let rank = 0;
-  let seconds = -1;
-  for (let i = 0; i < data.length; i += 1) {
-    if (data[i].seconds !== seconds) {
-      rank = i + 1;
-      seconds = data[i].seconds;
-    }
-    ranks.push({
-      ...data[i],
-      rank,
-    });
-  }
-
-  return ranks;
-};
+import { getRanks } from "../api/rank.js";
 
 export default class HitsujiRanking extends Phaser.Scene {
   constructor() {
